requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
httpx>=0.24.0
python-dotenv>=1.0.0
